+++ smartech_app/lib/app_inbox/utils/utils.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
+// import 'package:flutter_html/flutter_html.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
 showToast(String msg) {
@@ -26,7 +27,7 @@ Widget htmlText(String title) {
     children: [
       Html(
         data: title,
-        style: {"body": Style(margin: EdgeInsets.zero, padding: EdgeInsets.zero)},
+        style: {"body": Style(margin: Margins.zero, padding: HtmlPaddings.zero)},
       ),
       SizedBox(
         height: 4,