+++ smartech_app/lib/app_inbox/widgets/smt_simple_notification_view.dart
@@ -2,8 +2,6 @@ import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:smartech_app/app_inbox/utils/utils.dart';
-import 'package:smartech_app/deep_link_screen.dart';
-import 'package:smartech_app/navigator.dart';
 import 'package:smartech_appinbox/model/smt_appinbox_model.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -57,10 +55,10 @@ class SMTSimpleNotificationView extends StatelessWidget {
                                           Map<String, dynamic> dict = HashMap();
                                           dict["actionDeeplink"] = e.actionDeeplink;
                                           dict['isFromScreen'] = true;
-                                          NavigationUtilities.pushRoute(
-                                            DeepLinkScreen.route,
-                                            args: dict,
-                                          );
+                                          // NavigationUtilities.pushRoute(
+                                          //   DeepLinkScreen.route,
+                                          //   args: dict,
+                                          // );
                                         }
                                       },
                                       child: Text(