+++ smartech_app/lib/main.dart
@@ -1,4 +1,5 @@
 import 'dart:collection';
+import 'dart:convert';
 import 'dart:developer';
 import 'dart:io';
 import 'package:flutter/material.dart';
@@ -7,6 +8,7 @@ import 'package:smartech_app/deep_link_screen.dart';
 import 'package:smartech_app/events_utils.dart';
 import 'package:smartech_app/navigator.dart';
 import 'package:smartech_app/update_profile.dart';
+import 'package:smartech_appinbox/model/smt_appinbox_model.dart';
 import 'package:smartech_base/smartech_base.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'splash_screen.dart';
@@ -14,10 +16,22 @@ import 'splash_screen.dart';
 // DeepLinkNavigation deepLinkNavigation = DeepLinkNavigation();
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  Smartech().onHandleDeeplinkAction((String? link, Map<dynamic, dynamic>? map, bool? isAfterTerminated) async {
-    print("is after terminated " + isAfterTerminated.toString());
-    Future.delayed(const Duration(milliseconds: 500), () async {
-      Globle().deepLinkNavigation(link, map, isAfterTerminated);
+  // Smartech().onHandleDeeplinkAction((String? smtDeeplink, Map<dynamic, dynamic>? smtCustomPayload) async {
+  //   print("smtDeeplinkSource value :" + smtDeeplink.toString());
+  //   print("smtCustomPayload value :" + smtCustomPayload.toString());
+  //   Future.delayed(const Duration(milliseconds: 1000), () async {
+  //     Globle().deepLinkNavigation(smtDeeplink, smtCustomPayload);
+  //   });
+  // });
+
+  Smartech()
+      .onHandleDeeplink((String? smtDeeplinkSource, String? smtDeeplink, Map<dynamic, dynamic>? smtPayload, Map<dynamic, dynamic>? smtCustomPayload) async {
+    print("smtDeeplink value :" + smtDeeplink.toString());
+    print("smtCustomPayload value :" + smtCustomPayload.toString());
+    print("smtDeeplinkSource value :" + smtDeeplinkSource.toString());
+    print("smtPayload value :" + smtPayload.toString());
+    Future.delayed(const Duration(milliseconds: 1000), () async {
+      Globle().deepLinkNavigationWithPayload(smtDeeplinkSource, smtDeeplink, smtPayload, smtCustomPayload);
     });
   });
   await loadEventsJson();
@@ -36,7 +50,8 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
     super.initState();
 
     if (Platform.isAndroid) {
-      Smartech().setInAppCustomHTMLListener(customHTMLCallback);
+      // Smartech().setInAppCustomHTMLListener(customHTMLCallback);
+      Smartech().setOnInAppClickListener(onInAppCallback);
     }
   }
 
@@ -59,7 +74,18 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
   }
 
   Future<void> customHTMLCallback(Map<String, dynamic>? payload) async {
-    print(payload);
+    if (payload != null) {
+      log("customHTMLCallback payload" + payload.toString());
+    }
+  }
+
+  Future<void> onInAppCallback(String? deeplink, Map<String, dynamic>? payload) async {
+    if (deeplink != null) {
+      log("onInAppCallback deeplink" + deeplink.toString());
+    }
+    if (payload != null) {
+      log("onInAppCallback payload" + payload.toString());
+    }
   }
 
   @override
@@ -112,6 +138,7 @@ void getLocation() async {
 
 class Globle {
   static final Globle _singleton = Globle._internal();
+  SMTAppInboxMessage smtAppInboxMessage = SMTAppInboxMessage();
 
   factory Globle() {
     return _singleton;
@@ -120,24 +147,54 @@ class Globle {
   Globle._internal();
   late BuildContext context;
 
-  deepLinkNavigation(String? link, Map<dynamic, dynamic>? map, bool? isAfterTerminated) async {
-    if (link.toString() != "" || map!.isNotEmpty) {
+  deepLinkNavigation(String? link, Map<dynamic, dynamic>? payload) async {
+    if (link != null || payload!.isNotEmpty) {
       Map<String, dynamic> dict = HashMap();
-      log(map.toString());
-      dict["deepLinkData"] = map;
-      dict["deepLinkUrl"] = link;
-      dict["isFromScreen"] = false;
-      if (link!.contains("http")) {
+      log(payload.toString());
+      var encodedJson = json.encode(payload);
+      dict["smtCustomPayload"] = json.decode(encodedJson.toString());
+      dict["smtDeeplink"] = link;
+      if (dict["smtDeeplink"].contains("http")) {
         print("navigate to browser with url");
-        final Uri _url = Uri.parse(dict["deepLinkUrl"]);
+        final Uri _url = Uri.parse(dict["smtDeeplink"]);
         if (!await launchUrl(_url)) throw 'Could not launch $_url';
-      } else if (link.contains("smartechflutter://profile")) {
+      } else if (dict["smtDeeplink"].contains("smartechflutter://profile")) {
         NavigationUtilities.pushRoute(UpdateProfile.route);
       } else {
         NavigationUtilities.pushRoute(
           DeepLinkScreen.route,
           args: dict,
         );
+      }
+    } else {
+      return;
+    }
+  }
+
+  deepLinkNavigationWithPayload(
+      String? smtDeeplinkSource, String? smtDeeplink, Map<dynamic, dynamic>? smtPayload, Map<dynamic, dynamic>? smtCustomPayload) async {
+    if (smtPayload != null) {
+      smtAppInboxMessage = SMTAppInboxMessage.fromJson(smtPayload['smtPayload'] ?? smtPayload['data'] ?? smtPayload['payload'] ?? {});
+    }
+    print(smtAppInboxMessage);
+
+    if (smtDeeplink != null || smtCustomPayload!.isNotEmpty) {
+      Map<String, dynamic> dict = HashMap();
+      dict["smtDeeplink"] = smtDeeplink ?? "";
+      dict["smtCustomPayload"] = smtCustomPayload ?? {};
+      dict["smtDeeplinkSource"] = smtDeeplinkSource;
+      dict["smtPayload"] = smtPayload;
+      log(smtCustomPayload.toString());
+      if (dict["smtDeeplink"].contains("http")) {
+        print("navigate to browser with url");
+        final Uri _url = Uri.parse(dict["smtDeeplink"]);
+        print(_url);
+        if (!await launchUrl(_url)) throw 'Could not launch $_url';
+      } else if (dict["smtDeeplink"].contains("smartechflutter://profile")) {
+        NavigationUtilities.pushRoute(UpdateProfile.route);
+      } else {
+        print(dict);
+        NavigationUtilities.pushRoute(DeepLinkScreen.route, args: dict);
       }
     } else {
       return;