+++ smartech_appinbox/CHANGELOG.md
@@ -30,3 +30,7 @@
 ## 3.2.6
 
 * Updated gradle version.
+
+## 3.2.7
+
+* Updated model class according to Push Notification payload.